progress encapsed, hero name persists in wizard
