+++ js/wizard/steps/booking-details-step/booking-details-step.js
@@ -1,6 +1,18 @@
 export const bookingDetailsStep = () => {
    const bookingDetailsStepWrapper = document.createElement('div');
    bookingDetailsStepWrapper.innerText = 'Booking details';
+   //toptext
+   //progbar
+   //input{
+   //fighter
+   //name
+   //email
+   //}
+   //submit
+   //back
+
+   let topText = document.createElement('span');
+   let progressBar = document.createElement('img');
 
    return bookingDetailsStepWrapper;
 };
