+++ js/wizard/wizard.js
@@ -2,6 +2,7 @@ import { chooseFighterStep } from './steps/choose-fighter-step/choose-fighter-st
 import { bookingDetailsStep } from './steps/booking-details-step/booking-details-step.js';
 import { bookingConfirmationStep } from './steps/booking-confirmation-step/booking-confirmation-step.js';
 import { bigButtonMaker } from '../components/bigButtonMaker.js';
+import { progressMaker } from '../components/progressMaker.js';
 
 export const initWizard = (index = 0) => {
    let proceedButtonProps = {
@@ -20,6 +21,7 @@ export const initWizard = (index = 0) => {
    const maxSteps = steps.length;
 
    const wizardWrapper = document.createElement('div');
+   wizardWrapper.append(progressMaker({ index }));
    wizardWrapper.classList.add('genericScreenStyle');
    wizardWrapper.id = 'wizardWrapper';
 
@@ -29,6 +31,8 @@ export const initWizard = (index = 0) => {
    wizardWrapper.append(button);
 
    function nextWizardStep() {
+      let characterName = wizardWrapper.childNodes[1].getAttribute('characterName');
+      console.log(characterName);
       let wizardScreen = wizardWrapper.parentElement;
       wizardWrapper.remove();
       wizardScreen.append(initWizard((index + 1) % 3)); // modulo 3 in this step cicles back to choose-fighter-step
